src/tests/microservices/requirements.txt
# Copyright (C) 2024-2025 Intel Corporation
# SPDX-License-Identifier: Apache-2.0

pytest==7.4.4 # MIT
allure-pytest==2.13.5 # Apache 2.0
python-on-whales==0.73.0  # MIT
requests==2.32.3 # Apache 2.0
python-dotenv==1.0.1 # BSD License (BSD-3-Clause)
